Разделение конфигурации приложения
