+++ server/Config/Run.cs
@@ -1,0 +1,13 @@
+﻿namespace Rodnie.API.Config {
+    public static partial class Config {
+        public static WebApplication AppRun(this WebApplication app) {
+            app.UseAuthentication();
+            app.UseAuthorization();
+            app.MapControllers();
+
+            app.Run();
+
+            return app;
+        }
+    }
+}
